requirements.txt
django
django-crispy-forms
django-autocomplete-light
django-tables2
django-filter
